backend/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2==2.9.10
pydantic==2.10.4
more-itertools==10.5.0
tenacity==9.0.0
